+++ BraintreePaymentFlow/Public/ThreeDSecure/BTThreeDSecureResult.h
@@ -19,17 +19,17 @@
 /**
  True if the 3D Secure flow was successful
  */
-@property (nonatomic, assign) BOOL success DEPRECATED_MSG_ATTRIBUTE("Use the liabilityShiftPossible and liabilityShifted flags instead.");
+@property (nonatomic, assign) BOOL success DEPRECATED_MSG_ATTRIBUTE("Use `tokenizedCard.threeDSecureInfo.liabilityShiftPossible` and `tokenizedCard.threeDSecureInfo.liabilityShifted` instead.");
 
 /**
  Indicates whether the liability for fraud has been shifted away from the merchant
  */
-@property (nonatomic, assign) BOOL liabilityShifted;
+@property (nonatomic, assign) BOOL liabilityShifted DEPRECATED_MSG_ATTRIBUTE("Use `tokenizedCard.threeDSecureInfo.liabilityShifted` instead.");
 
 /**
  Indicates whether the card was eligible for 3D Secure
  */
-@property (nonatomic, assign) BOOL liabilityShiftPossible;
+@property (nonatomic, assign) BOOL liabilityShiftPossible DEPRECATED_MSG_ATTRIBUTE("Use `tokenizedCard.threeDSecureInfo.liabilityShiftPossible` instead.");
 
 /**
  The `BTCardNonce` resulting from the 3D Secure flow
@@ -39,7 +39,7 @@
 /**
  The error message when the 3D Secure flow is unsuccessful
  */
-@property (nonatomic, copy) NSString *errorMessage;
+@property (nonatomic, copy) NSString *errorMessage DEPRECATED_MSG_ATTRIBUTE("Use `tokenizedCard.threeDSecureInfo.errorMessage` instead.");
 
 /**
  Initialize a BTThreeDSecureResult